+++ src/components/models/loaded_model.jl
