+++ src/components/serialization/serialization_mdarrays.jl
@@ -1,5 +1,3 @@
-module Serialization
-
 # Preference based serialization of multidimensional arrays, such as matrices and tensors.
 # The major problem and potential solution is described here:
 # https://github.com/lazydynamics/RxInferServer/issues/59
@@ -248,121 +246,6 @@ function from_string(preference::AbstractString)
 end
 end
 
-"""
-    UnsupportedPreferenceError(option, options)
-
-Error thrown when an unknown `options` value is used for a given set of `options`.
-"""
-struct UnsupportedPreferenceError{T, P} <: Exception
-    option::T
-    options::P
-end
-
-function Base.showerror(io::IO, e::UnsupportedPreferenceError)
-    print(io, "unknown preference `$(e.option)` for `$(e.options.OptionName)`.")
-    print(io, "Available preferences are ")
-    join(io, map(s -> string('`', e.options.to_string(s), '`'), e.options.AvailableOptions), ", ", " and ")
-end
-
-# Using `JSON` instead for `to_json` of `JSON3` here is intentional.
-# `JSON3` does not support custom serializers, and `JSON` does.
-import JSON
-import JSON.Writer:
-    StructuralContext,
-    begin_object,
-    end_object,
-    begin_array,
-    end_array,
-    show_pair,
-    show_key,
-    show_json,
-    show_element,
-    delimit,
-    indent
-
-"""
-    JSONSerialization(; kwargs...)
-
-Type-safe JSON serializer for OpenAPI data types with configurable multi-dimensional array handling.
-
-# Keywords
-- `mdarray_repr`: Multi-dimensional array representation format, see [`RxInferServer.Serialization.MultiDimensionalArrayRepr`](@ref)
-- `mdarray_data`: Multi-dimensional array data encoding format, see [`RxInferServer.Serialization.MultiDimensionalArrayData`](@ref)
-"""
-Base.@kwdef struct JSONSerialization <: JSON.Serializations.Serialization
-    mdarray_repr::UInt8 = MultiDimensionalArrayRepr.Dict
-    mdarray_data::UInt8 = MultiDimensionalArrayData.ArrayOfArrays
-end
-
-struct UnsupportedTypeSerializationError <: Exception
-    type::Type
-end
-
-function Base.showerror(io::IO, e::UnsupportedTypeSerializationError)
-    print(io, "serialization of type $(e.type) is not supported")
-end
-
-function show_json(::StructuralContext, ::JSONSerialization, value)
-    throw(UnsupportedTypeSerializationError(typeof(value)))
-end
-
-# OpenAPI has the following types defined
-# https://swagger.io/docs/specification/v3_0/data-models/data-types/
-# - string (this includes dates and files)
-# - number 
-# - integer
-# - boolean
-# - array 
-# - object
-show_json(io::StructuralContext, ::JSONSerialization, value::String) =
-    show_json(io, JSON.StandardSerialization(), value)
-show_json(io::StructuralContext, ::JSONSerialization, value::Number) =
-    show_json(io, JSON.StandardSerialization(), value)
-show_json(io::StructuralContext, ::JSONSerialization, value::Bool) = show_json(io, JSON.StandardSerialization(), value)
-show_json(io::StructuralContext, ::JSONSerialization, value::Union{Missing, Nothing}) =
-    show_json(io, JSON.StandardSerialization(), value)
-
-# String - date-time including timezones
-using Dates, TimeZones
-
-show_json(io::StructuralContext, ::JSONSerialization, value::DateTime) =
-    show_json(io, JSON.StandardSerialization(), value)
-show_json(io::StructuralContext, ::JSONSerialization, value::ZonedDateTime) =
-    show_json(io, JSON.StandardSerialization(), value)
-
-# Vector-like values
-show_json(io::StructuralContext, s::JSONSerialization, value::Tuple) = __json_serialization_vector_like(io, s, value)
-show_json(io::StructuralContext, s::JSONSerialization, value::AbstractVector) =
-    __json_serialization_vector_like(io, s, value)
-
-function __json_serialization_vector_like(io::StructuralContext, s::JSONSerialization, vectorlike)
-    begin_array(io)
-    foreach(element -> show_element(io, s, element), vectorlike)
-    end_array(io)
-end
-
-# Dict-like values
-show_json(io::StructuralContext, s::JSONSerialization, value::NamedTuple) = __json_serialization_dict_like(io, s, value)
-show_json(io::StructuralContext, s::JSONSerialization, value::AbstractDict) =
-    __json_serialization_dict_like(io, s, value)
-
-function __json_serialization_dict_like(io::StructuralContext, s::JSONSerialization, dictlike)
-    begin_object(io)
-    foreach(pair -> show_pair(io, s, pair), pairs(dictlike))
-    end_object(io)
-end
-
-# We also support serialization of OpenAPI defined types, which are subtypes of `APIModel`
-using RxInferServerOpenAPI
-
-function show_json(io::StructuralContext, s::JSONSerialization, value::RxInferServerOpenAPI.OpenAPI.APIModel)
-    begin_object(io)
-    for field in propertynames(value)
-        show_pair(io, s, field => getproperty(value, field))
-    end
-    end_object(io)
-end
-
 # Multi-dimensional arrays, preference based serialization
 
 function show_json(io::StructuralContext, serialization::JSONSerialization, value::AbstractArray)
@@ -493,44 +376,4 @@ end
 ## MultiDimensionalArrayData.None implementation
 function __show_mdarray_data_none(io::StructuralContext, serialization::JSONSerialization, array::AbstractArray)
     show_json(io, serialization, nothing)
-end
-
-"""
-    to_json([io::IO], [serialization::JSONSerialization], value)
-
-Serialize `value` to JSON using the specified serialization strategy.
-
-Returns a string if `io` is not provided.
-
-See also: [`RxInferServer.Serialization.JSONSerialization`](@ref)
-"""
-function to_json end
-
-to_json(value) = to_json(JSONSerialization(), value)
-to_json(s::JSONSerialization, value) = sprint(show_json, s, value)
-to_json(io::IO, value) = show_json(io, JSONSerialization(), value)
-to_json(io::IO, s::JSONSerialization, value) = show_json(io, s, value)
-
-"""
-    from_json(string)
-
-Parse a JSON string into Julia data structures.
-
-```jldoctest
-julia> import RxInferServer.Serialization: JSONSerialization, MultiDimensionalArrayRepr, MultiDimensionalArrayData, to_json, from_json
-
-julia> s = JSONSerialization(mdarray_repr = MultiDimensionalArrayRepr.Dict, mdarray_data = MultiDimensionalArrayData.ArrayOfArrays);
-
-julia> from_json(to_json(s, [1 2; 3 4]))
-Dict{String, Any} with 4 entries:
-  "shape"    => Any[2, 2]
-  "encoding" => "array_of_arrays"
-  "data"     => Any[Any[1, 2], Any[3, 4]]
-  "type"     => "mdarray"
-```
-
-Note: No post-processing is performed on the deserialized value.
-"""
-from_json(string) = JSON.parse(string)
-
 end
